+++ cases/credit_scoring/credit_scoring_problem_multiobj.py
@@ -5,7 +5,8 @@ import numpy as np
 from sklearn.metrics import roc_auc_score as roc_auc
 
 from cases.credit_scoring.credit_scoring_problem import get_scoring_data
-from fedot.core.composer.gp_composer.gp_composer import GPComposerBuilder, PipelineComposerRequirements
+from fedot.core.composer.composer_builder import ComposerBuilder
+from fedot.core.composer.gp_composer.gp_composer import PipelineComposerRequirements
 from fedot.core.data.data import InputData
 from fedot.core.optimisers.gp_comp.gp_optimiser import GPGraphOptimiserParameters, GeneticSchemeTypesEnum
 from fedot.core.optimisers.gp_comp.operators.selection import SelectionTypesEnum
@@ -65,7 +66,7 @@ def run_credit_scoring_problem(train_file_path, test_file_path,
                                                       selection_types=[SelectionTypesEnum.spea2])
 
     # Create builder for composer and set composer params
-    builder = GPComposerBuilder(task=task).with_requirements(composer_requirements).with_metrics(
+    builder = ComposerBuilder(task=task).with_requirements(composer_requirements).with_metrics(
         metrics).with_optimiser(parameters=optimiser_parameters)
 
     # Create GP-based composer
