+++ fedot/core/composer/composer.py
@@ -48,7 +48,7 @@ class ComposerRequirements:
 class Composer(ABC):
     """
     Base class used for receiving composite operations via optimization
-    :param optimiser: optimiser generated in GPComposerBuilder
+    :param optimiser: optimiser generated in ComposerBuilder
     :param metrics: metrics used to define the quality of found solution.
     :param composer_requirements: requirements for composition process
     :param initial_pipeline: defines the initial state of the population. If None then initial population is random.
