+++ fedot/core/composer/gp_composer/gp_composer.py
@@ -5,29 +5,19 @@ from functools import partial
 from multiprocessing import set_start_method
 from typing import Any, Callable, List, Optional, Tuple, Union
 
-from deap import tools
-
-from fedot.core.composer.advisor import PipelineChangeAdvisor
 from fedot.core.composer.cache import OperationsCache
 from fedot.core.composer.composer import Composer, ComposerRequirements
-from fedot.core.composer.gp_composer.specific_operators import boosting_mutation, parameter_change_mutation
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.data.multi_modal import MultiModalData
 from fedot.core.log import Log, default_log
-from fedot.core.optimisers.adapters import PipelineAdapter
-from fedot.core.optimisers.gp_comp.gp_optimiser import EvoGraphOptimiser, GPGraphOptimiserParameters
-from fedot.core.optimisers.gp_comp.operators.inheritance import GeneticSchemeTypesEnum
-from fedot.core.optimisers.gp_comp.operators.mutation import MutationStrengthEnum, MutationTypesEnum
-from fedot.core.optimisers.gp_comp.operators.regularization import RegularizationTypesEnum
-from fedot.core.optimisers.gp_comp.param_free_gp_optimiser import EvoGraphParameterFreeOptimiser
-from fedot.core.optimisers.optimizer import GraphGenerationParams, GraphOptimiser, GraphOptimiserParameters
+from fedot.core.optimisers.gp_comp.operators.mutation import MutationStrengthEnum
+from fedot.core.optimisers.graph import OptGraph
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.pipelines.validation import common_rules, ts_rules, validate
-from fedot.core.repository.operation_types_repository import get_operations_for_task
-from fedot.core.repository.quality_metrics_repository import (ClassificationMetricsEnum, MetricsEnum,
-                                                              MetricsRepository, RegressionMetricsEnum)
-from fedot.core.repository.tasks import Task, TaskTypesEnum
+from fedot.core.repository.quality_metrics_repository import (MetricsEnum,
+                                                              MetricsRepository)
+from fedot.core.repository.tasks import TaskTypesEnum
 from fedot.core.validation.compose.tabular import table_metric_calculation
 from fedot.core.validation.compose.time_series import ts_metric_calculation
 from fedot.remote.remote_evaluator import RemoteEvaluator, init_data_for_remote_execution
@@ -70,7 +60,7 @@ class PipelineComposerRequirements(ComposerRequirements):
 class GPComposer(Composer):
     """
     Genetic programming based composer
-    :param optimiser: optimiser generated in GPComposerBuilder
+    :param optimiser: optimiser generated in ComposerBuilder
     :param metrics: metrics used to define the quality of found solution.
     :param composer_requirements: requirements for composition process
     :param initial_pipeline: defines the initial state of the population. If None then initial population is random.
@@ -145,15 +135,18 @@ class GPComposer(Composer):
 
         opt_result = self.optimiser.optimise(objective_function_for_pipeline,
                                              on_next_iteration_callback=on_next_iteration_callback)
-        best_pipeline = [self.optimiser.graph_generation_params.adapter.restore(graph)
-                         for graph in opt_result] if isinstance(opt_result, list) \
-            else self.optimiser.graph_generation_params.adapter.restore(opt_result)
+        best_pipeline = self._convert_opt_results_to_pipeline(opt_result)
 
         self.log.info('GP composition finished')
         self.cache.clear()
         if is_tune:
             self.tune_pipeline(best_pipeline, data, self.composer_requirements.timeout)
         return best_pipeline
+
+    def _convert_opt_results_to_pipeline(self, opt_result: Union[OptGraph, List[OptGraph]]) -> Pipeline:
+        return [self.optimiser.graph_generation_params.adapter.restore(graph)
+                for graph in opt_result] if isinstance(opt_result, list) \
+            else self.optimiser.graph_generation_params.adapter.restore(opt_result)
 
     def _cv_validation_metric_build(self, data):
         """ Prepare function for metric evaluation based on task """
@@ -229,93 +222,3 @@ class GPComposer(Composer):
     @property
     def history(self):
         return self.optimiser.history
-
-
-class GPComposerBuilder:
-    def __init__(self, task: Task):
-        self._composer = GPComposer()
-        self.optimiser = EvoGraphOptimiser
-        self.optimiser_parameters = GPGraphOptimiserParameters()
-        self.task = task
-        self.set_default_composer_params()
-
-    def with_optimiser(self, optimiser: Optional[GraphOptimiser] = None,
-                       parameters: Optional[GraphOptimiserParameters] = None):
-        if optimiser is not None:
-            self.optimiser = optimiser
-        if parameters is not None:
-            self.optimiser_parameters = parameters
-        return self
-
-    def with_requirements(self, requirements: PipelineComposerRequirements):
-        self._composer.composer_requirements = requirements
-        return self
-
-    def with_metrics(self, metrics: Union[List[MetricsEnum], MetricsEnum]):
-        if type(metrics) is not list:
-            metrics = [metrics]
-        self._composer.metrics = metrics
-        return self
-
-    def with_initial_pipeline(self, initial_pipeline: Optional[Pipeline]):
-        self._composer.initial_pipeline = initial_pipeline
-        return self
-
-    def with_logger(self, logger):
-        self._composer.log = logger
-        return self
-
-    def with_cache(self, cache_path: str = None, use_existing=False):
-        self._composer.cache_path = cache_path
-        self._composer.use_existing_cache = use_existing
-        return self
-
-    def set_default_composer_params(self):
-        """ Method set metrics and composer requirements """
-        if not self._composer.composer_requirements:
-            # Get all available operations for task
-            operations = get_operations_for_task(task=self.task, mode='all')
-
-            # Set protected attributes to composer
-            self._composer.composer_requirements = PipelineComposerRequirements(primary=operations,
-                                                                                secondary=operations)
-        if not self._composer.metrics:
-            metric_function = ClassificationMetricsEnum.ROCAUC_penalty
-            if self.task.task_type in (TaskTypesEnum.regression, TaskTypesEnum.ts_forecasting):
-                metric_function = RegressionMetricsEnum.RMSE
-
-            # Set metric
-            self._composer.metrics = [metric_function]
-
-    def build(self) -> Composer:
-        optimiser_type = self.optimiser
-        if (optimiser_type == EvoGraphOptimiser and
-                self.optimiser_parameters.genetic_scheme_type == GeneticSchemeTypesEnum.parameter_free):
-            optimiser_type = EvoGraphParameterFreeOptimiser
-
-        graph_generation_params = GraphGenerationParams(adapter=PipelineAdapter(self._composer.log),
-                                                        advisor=PipelineChangeAdvisor())
-
-        archive_type = None
-        if len(self._composer.metrics) > 1:
-            self.optimiser_parameters.archive_type = tools.ParetoFront()
-            # TODO add possibility of using regularization in MO alg
-            self.optimiser_parameters.regularization_type = RegularizationTypesEnum.none
-            self.optimiser_parameters.multi_objective = True
-
-        if self.optimiser_parameters.mutation_types is None:
-            self.optimiser_parameters.mutation_types = [boosting_mutation, parameter_change_mutation,
-                                                        MutationTypesEnum.single_edge,
-                                                        MutationTypesEnum.single_change,
-                                                        MutationTypesEnum.single_drop,
-                                                        MutationTypesEnum.single_add]
-
-        optimiser = optimiser_type(initial_graph=self._composer.initial_pipeline,
-                                   requirements=self._composer.composer_requirements,
-                                   graph_generation_params=graph_generation_params,
-                                   parameters=self.optimiser_parameters, log=self._composer.log,
-                                   metrics=self._composer.metrics)
-
-        self._composer.optimiser = optimiser
-
-        return self._composer
